:hammer: Refactor publishers
