+++ ern-container-gen/src/index.js
@@ -4,14 +4,12 @@ import { generateMiniAppsComposite as _generateMiniAppsComposite } from './utils
 import _generateContainer from './generateContainer'
 import _IosGenerator from './generators/ios/IosGenerator'
 import _AndroidGenerator from './generators/android/AndroidGenerator'
-import _Publisher from './publishers/Publisher'
 import _ContainerGeneratorConfig from './ContainerGeneratorConfig'
 
 export const AndroidGenerator = _AndroidGenerator
 export const IosGenerator = _IosGenerator
 export const generateContainer = _generateContainer
 export const generateMiniAppsComposite = _generateMiniAppsComposite
-export const Publisher = _Publisher
 export const ContainerGeneratorConfig = _ContainerGeneratorConfig
 
 export default ({
@@ -19,6 +17,5 @@ export default ({
   IosGenerator: _IosGenerator,
   generateContainer: _generateContainer,
   generateMiniAppsComposite: _generateMiniAppsComposite,
-  Publisher: _Publisher,
   ContainerGeneratorConfig: _ContainerGeneratorConfig
 })
