--- ern-container-gen/src/publishers/Publisher.js
@@ -1,6 +1,0 @@
-// @flow
-export interface Publisher {
-  publish(any): any;
-  + name: string;
-  + url: string
-}
